+++ pNeuma_simulator/__init__.py
@@ -1,0 +1,10 @@
+# from .animations import draw, ring  # noqa F401
+# from .batch import batch  # noqa F401
+# from .identify import identify  # noqa F401
+# from .infront import infront  # noqa F401
+# from .main import main  # noqa F401
+# from .newton_iteration import newton_iteration  # noqa F401
+# from .results import loader, aggregate, intersect  # noqa F401
+# from .zipdir import zipdir  # noqa F401
+
+__version__ = "0.0.0"
