+++ pNeuma_simulator/dca/__init__.py
@@ -1,3 +1,3 @@
-from dca.dca import calc_dtc, ellipses  # noqa F401
+from .dca import calc_dtc, ellipses  # noqa F401
 
 __version__ = "0.0.0"
