+++ pNeuma_simulator/initialization/__init__.py
@@ -1,6 +1,6 @@
-from initialization.initialization import budget, f, vo, ov  # noqa F401
-from initialization.equilibrium import equilibrium  # noqa F401
-from initialization.poissondic import PoissonDic  # noqa F401
-from initialization.synthetic_fd import synthetic_fd  # noqa F401
+from .initialization import budget, f, vo, ov  # noqa F401
+from .equilibrium import equilibrium  # noqa F401
+from .poissondisc import PoissonDisc  # noqa F401
+from .synthetic_fd import synthetic_fd  # noqa F401
 
 __version__ = "0.0.0"
